Completely rewrote main formula and updated README
Started with a whole new calculation formula that is now
able to go through all single and double digit combinations,
and I just have to figure out how to expand that to all possibilities.

Also updated the README to reflect that I wanted this as more a journal
of my progress in solving this issue.