Iteratively goes through every number combination
Figured out a recursive function in order to go through every
combination of digits (I'm pretty sure?), so now it's a matter
of figuring out how to compare that to the sum I'm looking for and
adding to a dictionary. But it seems like so much work for the computer
to go through every combination 45 times (45 being the largest sum
you can reach with nine digits).

Having just written that I think the goal would be to have a series of
sums in a dictionary, and each time I go through add that combination
to that sum to only go through every iteration once instead of 45 times.