Solved recursion, having a dictionary problem
I'm able to go through iteratively every combination of numbers and
then compare them to a sum and add them to the dictionary, but I can't
add them as a separate list to that dictionary entry for some reason,
it either mashes all the lists into one or it shows empty lists.

For example for the sum 5, I either get a series of empty lists, or I
get a long list of all combinations together: [1, 4, 2, 3, 3, 2, 4, 1],
but I want to see each individual list as the variables that total that
sum. Once I figure that out, I can find a way to compare the current
contents and sort the lists so that I don't add duplicate possible
combinations.